# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/chats": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый чат (приватный или групповой) или присоединяет пользователя к существующему приватному чату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Создание чата",
                "parameters": [
                    {
                        "description": "Данные для создания чата",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.CreateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Чат успешно создан или пользователь присоединен к существующему",
                        "schema": {
                            "$ref": "#/definitions/routes.CreateChatResponse"
                        }
                    }
                }
            }
        },
        "/auth/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Получение друзей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Создание друга",
                "parameters": [
                    {
                        "description": "Данные",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.FriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.FriendStatus"
                        }
                    }
                }
            }
        },
        "/auth/friends/accepted": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Принять друга",
                "parameters": [
                    {
                        "description": "Принять друга",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.AcceptedFriendRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "routes.AcceptedFriendRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "string"
                }
            }
        },
        "routes.CreateChatRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type_chat": {
                    "$ref": "#/definitions/routes.TypeChat"
                }
            }
        },
        "routes.CreateChatResponse": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "created": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "routes.FriendRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "routes.FriendStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusAccepted"
            ]
        },
...
        "routes.TypeChat": {
            "type": "string",
            "enum": [
                "private",
                "group"
            ],
            "x-enum-varnames": [
                "TypeChatPrivate",
                "TypeChatGroup"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите ваш токен",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}